The switch between the legacy version and the localized version (lts) of the topological simplification is now done in the ttk layer and no longer in the vtk layer. TopologicalSimplification is now only a container class to call the chosen implementation.
