+++ core/base/legacyTopologicalSimplification/CMakeLists.txt
@@ -1,0 +1,8 @@
+ttk_add_base_library(legacyTopologicalSimplification
+  SOURCES
+    LegacyTopologicalSimplification.cpp
+  HEADERS
+    LegacyTopologicalSimplification.h
+  DEPENDS
+    triangulation
+    )
