+++ core/base/legacyTopologicalSimplification/LegacyTopologicalSimplification.cpp
@@ -1,0 +1,5 @@
+#include <LegacyTopologicalSimplification.h>
+
+ttk::LegacyTopologicalSimplification::LegacyTopologicalSimplification() {
+  this->setDebugMsgPrefix("LegacyTopologicalSimplification");
+}
