+++ core/base/topologicalSimplification/CMakeLists.txt
@@ -5,4 +5,6 @@ ttk_add_base_library(topologicalSimplification
     TopologicalSimplification.h
   DEPENDS
     triangulation
+    legacyTopologicalSimplification
+    localizedTopologicalSimplification
     )
